feat(http-core): refactor requestMethod decorator and corresponding tests
