+++ packages/http/libraries/core/src/http/decorators/All.spec.ts
@@ -1,10 +1,10 @@
 import { afterAll, beforeAll, describe, expect, it, vitest } from 'vitest';
 
-vitest.mock('./RequestMethod');
+vitest.mock('../calculations/RequestMethod');
 
+import { requestMethod } from '../calculations/requestMethod';
 import { RequestMethodType } from '../models/RequestMethodType';
 import { ALL } from './All';
-import { requestMethod } from './RequestMethod';
 
 describe(ALL.name, () => {
   describe('when called', () => {
