+++ packages/http/libraries/core/src/http/decorators/Delete.ts
@@ -1,5 +1,5 @@
+import { requestMethod } from '../calculations/requestMethod';
 import { RequestMethodType } from '../models/RequestMethodType';
-import { requestMethod } from './RequestMethod';
 
 export const DELETE: (path?: string) => MethodDecorator = (
   path?: string,
