+++ packages/http/libraries/core/src/http/decorators/Head.ts
@@ -1,5 +1,5 @@
+import { requestMethod } from '../calculations/requestMethod';
 import { RequestMethodType } from '../models/RequestMethodType';
-import { requestMethod } from './RequestMethod';
 
 export const HEAD: (path?: string) => MethodDecorator = (
   path?: string,
