+++ packages/http/libraries/core/src/http/decorators/Patch.ts
@@ -1,5 +1,5 @@
+import { requestMethod } from '../calculations/requestMethod';
 import { RequestMethodType } from '../models/RequestMethodType';
-import { requestMethod } from './RequestMethod';
 
 export const PATCH: (path?: string) => MethodDecorator = (
   path?: string,
